Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-encrypt.txt

Encrypted OK with 64 bytes
